test: cleanup testClusterWorkerSuite.
